feat: 노트 미리보기 모달창 구현 중..., 파일 분리
