+++ src/pages/Library/components/NoteModal.jsx
@@ -1,0 +1,61 @@
+import styled from 'styled-components';
+
+const ModalBackdrop = styled.div`
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background: rgba(0, 0, 0, 0.5);
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  z-index: 1000;
+`;
+
+const ModalContent = styled.div`
+  background: #fff;
+  padding: 2rem;
+  border-radius: 8px;
+  width: 60%;
+  max-width: 800px;
+  max-height: 80%;
+  overflow-y: auto;
+  position: relative;
+`;
+
+const CloseButton = styled.button`
+  position: absolute;
+  top: 1rem;
+  right: 1rem;
+  background: none;
+  border: none;
+  font-size: 1.5rem;
+  cursor: pointer;
+`;
+
+const ModalTitle = styled.h2`
+  font-size: 1.5rem;
+  margin-bottom: 1rem;
+`;
+
+const ModalBody = styled.div`
+  font-size: 1rem;
+  line-height: 1.5;
+`;
+
+const NoteModal = ({ show, onClose, title, content }) => {
+  if (!show) return null;
+
+  return (
+    <ModalBackdrop onClick={onClose}>
+      <ModalContent onClick={e => e.stopPropagation()}>
+        <CloseButton onClick={onClose}>&times;</CloseButton>
+        <ModalTitle>{title}</ModalTitle>
+        <ModalBody>{content}</ModalBody>
+      </ModalContent>
+    </ModalBackdrop>
+  );
+};
+
+export default NoteModal;
